Explicitly create KubeVirt service accounts token if not found
K8s 1.24 disabled automatic token creation for service account,
thus it is required to create them by explicitly for CSI access
