+++ pkg/provider/cloud/kubevirt/data_volume.go
@@ -28,99 +28,15 @@ import (
 	"k8c.io/kubermatic/v2/pkg/resources/reconciling"
 
 	corev1 "k8s.io/api/core/v1"
-	rbacv1 "k8s.io/api/rbac/v1"
 	storagev1 "k8s.io/api/storage/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/sets"
-	restclient "k8s.io/client-go/rest"
 	utilpointer "k8s.io/utils/pointer"
 	ctrlruntimeclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-var (
-	csiServiceAccountNamespace = metav1.NamespaceDefault
-	csiResourceName            = "kubevirt-csi"
-)
-
-const InfraStorageClassAnnotation = "kubevirt-initialization.k8c.io/initialize-sc"
-
 type StorageClassAnnotationFilter func(map[string]string) bool
-
-func csiServiceAccountCreator(name string) reconciling.NamedServiceAccountCreatorGetter {
-	return func() (string, reconciling.ServiceAccountCreator) {
-		return name, func(sa *corev1.ServiceAccount) (*corev1.ServiceAccount, error) {
-			return sa, nil
-		}
-	}
-}
-
-func csiRoleCreator(name string) reconciling.NamedRoleCreatorGetter {
-	return func() (string, reconciling.RoleCreator) {
-		return name, func(r *rbacv1.Role) (*rbacv1.Role, error) {
-			r.Rules = []rbacv1.PolicyRule{
-				{
-					APIGroups: []string{"cdi.kubevirt.io"},
-					Resources: []string{"datavolumes"},
-					Verbs:     []string{"get", "create", "delete"},
-				},
-				{
-					APIGroups: []string{"kubevirt.io"},
-					Resources: []string{"virtualmachineinstances"},
-					Verbs:     []string{"list"},
-				},
-				{
-					APIGroups: []string{"subresources.kubevirt.io"},
-					Resources: []string{"virtualmachineinstances/addvolume", "virtualmachineinstances/removevolume"},
-					Verbs:     []string{"update"},
-				},
-			}
-
-			return r, nil
-		}
-	}
-}
-
-func csiRoleBindingCreator(name, namespace string) reconciling.NamedRoleBindingCreatorGetter {
-	return func() (string, reconciling.RoleBindingCreator) {
-		return name, func(rb *rbacv1.RoleBinding) (*rbacv1.RoleBinding, error) {
-			rb.Subjects = []rbacv1.Subject{
-				{
-					Kind:      "ServiceAccount",
-					Name:      name,
-					Namespace: namespace,
-				},
-			}
-
-			rb.RoleRef = rbacv1.RoleRef{
-				APIGroup: "rbac.authorization.k8s.io",
-				Kind:     "Role",
-				Name:     name,
-			}
-
-			return rb, nil
-		}
-	}
-}
-
-// reconcileCSIRoleRoleBinding reconciles the Role and Rolebindings needed by CSI driver.
-func reconcileCSIRoleRoleBinding(ctx context.Context, namespace string, client ctrlruntimeclient.Client, restConfig *restclient.Config) error {
-	roleCreators := []reconciling.NamedRoleCreatorGetter{
-		csiRoleCreator(csiResourceName),
-	}
-	if err := reconciling.ReconcileRoles(ctx, roleCreators, namespace, client); err != nil {
-		return err
-	}
-
-	roleBindingCreators := []reconciling.NamedRoleBindingCreatorGetter{
-		csiRoleBindingCreator(csiResourceName, csiServiceAccountNamespace),
-	}
-	if err := reconciling.ReconcileRoleBindings(ctx, roleBindingCreators, namespace, client); err != nil {
-		return err
-	}
-
-	return nil
-}
 
 func dataVolumeCreator(datavolume *cdiv1beta1.DataVolume) reconciling.NamedCDIv1beta1DataVolumeCreatorGetter {
 	return func() (name string, create reconciling.CDIv1beta1DataVolumeCreator) {
